+++ framework/boot.cpp
@@ -20,6 +20,8 @@
 
 #include <foundation/foundation.h>
 
+#undef ERROR_ACCESS_DENIED
+
 #if BUILD_DEVELOPMENT
 static bool _run_tests = false;
 #endif
@@ -71,6 +73,12 @@ FOUNDATION_STATIC void main_handle_debug_break()
 FOUNDATION_STATIC int main_process_command_line(foundation_config_t& config, application_t& application)
 {
     LOG_PREFIX(false);
+
+    if (BUILD_DEPLOY && environment_argument("run-tests"))
+    {
+        log_error(0, ERROR_ACCESS_DENIED, STRING_CONST("Tests cannot run in deploy mode"));
+        return -1;
+    }
 
     if (environment_argument("version"))
     {