+++ sources/tests/example_tests.cpp
@@ -9,18 +9,11 @@
  
 #include <framework/tests/test_utils.h>
 
-#include <framework/common.h>
-#include <framework/search_query.h>
-#include <framework/search_database.h>
-
-#include <foundation/random.h>
-
-#include <doctest/doctest.h>
-
 TEST_SUITE("Examples")
 {
     TEST_CASE("Example #1")
     {
+        CHECK(environment_argument("run-tests"));
     }
 }